+++ R/states.R
@@ -48,7 +48,7 @@
 #' * locality_id - the name of locality (in some cases identical to logger id, see [mc_read_files])
 #' * logger_index - index of logger in myClim object at the locality. See [mc_info_logger].   
 #' * sensor_name - sensor name either original (e.g., TMS_T1, T_C), or calculated/renamed (e.g., "TMS_T1_max", "my_sensor01") 
-#' * tag - category of state (e.g., "error", "source", "quality")  
+#' * tag - category of state (e.g., "conflict", "error", "source", "quality")  
 #' * start - start datetime
 #' * end - end datetime
 #' * value - value of tag (e.g., "out of soil", "c:/users/John/tmsData/data_911235678.csv")
@@ -100,7 +100,7 @@ mc_states_insert <- function(data, states_table) {
 #' * locality_id - the name of locality (in some cases identical to logger id, see details of [mc_read_files])
 #' * logger_index - index of logger in myClim object at the locality. See [mc_info_logger].
 #' * sensor_name - sensor name either original (e.g., TMS_T1, T_C), or calculated/renamed (e.g., "TMS_T1_max", "my_sensor01") 
-#' * tag - category of state (e.g., "error", "source", "quality")  
+#' * tag - category of state (e.g., "conflict", "error", "source", "quality")  
 #' * start - start datetime
 #' * end - end datetime
 #' * value - value of tag (e.g., "out of soil", "c:/users/John/tmsData/data_911235678.csv")
@@ -124,7 +124,7 @@ mc_states_update <- function(data, states_table) {
     return(states_table)
 }
 
-.states_run <- function(data, states_table, action_function) {
+.states_run <- function(data, states_table, action_function, edit_datetimes=TRUE) {
     is_agg <-.common_is_agg_format(data)
     data_env <- new.env()
     data_env$data <- data
@@ -149,7 +149,7 @@ mc_states_update <- function(data, states_table) {
             return()
         }
         states_table <- dplyr::select(.x, "tag", "start", "end", "value")
-        action_function(data_env, locality_id, logger_index, sensor_name, states_table)
+        action_function(data_env, locality_id, logger_index, sensor_name, states_table, edit_datetimes)
     }
 
     sensor_prep_function <- function(.x, .y) {
@@ -264,8 +264,11 @@ mc_states_update <- function(data, states_table) {
     return(states_table)
 }
 
-.states_insert <- function(data_env, locality_id, logger_index, sensor_name, states_table) {
-    states_table <- .states_edit_datetimes(data_env$data, locality_id, logger_index, states_table)
+.states_insert <- function(data_env, locality_id, logger_index, sensor_name, states_table, edit_datetimes) {
+    if(edit_datetimes) {
+        states_table <- .states_edit_datetimes(data_env$data, locality_id, logger_index, states_table)
+    }
+    states_table <- as.data.frame(states_table)
     if(.common_is_agg_format(data_env$data)) {
         data_env$data$localities[[locality_id]]$sensors[[sensor_name]]$states <-
             dplyr::bind_rows(data_env$data$localities[[locality_id]]$sensors[[sensor_name]]$states,
@@ -277,8 +280,11 @@ mc_states_update <- function(data, states_table) {
     }
 }
 
-.states_update <- function(data_env, locality_id, logger_index, sensor_name, states_table) {
-    states_table <- .states_edit_datetimes(data_env$data, locality_id, logger_index, states_table)
+.states_update <- function(data_env, locality_id, logger_index, sensor_name, states_table, edit_datetimes) {
+    if(edit_datetimes) {
+        states_table <- .states_edit_datetimes(data_env$data, locality_id, logger_index, states_table)
+    }
+    states_table <- as.data.frame(states_table)
     if(.common_is_agg_format(data_env$data)) {
         data_env$data$localities[[locality_id]]$sensors[[sensor_name]]$states <- states_table
     } else {
@@ -287,15 +293,27 @@ mc_states_update <- function(data, states_table) {
 }
 
 .states_edit_datetimes <- function(data, locality_id, logger_index, states_table) {
+    period <- NULL
+    step <- NULL
+    is_agg <- .common_is_agg_format(data)
     date_interval <- .states_get_item_range(data, locality_id, logger_index)
-    period <- .common_get_period_from_data(data, locality_id, logger_index)
+    if(is_agg){
+        period <- .common_get_period_from_data(data, locality_id, logger_index)
+    } else {
+        step <- data$localities[[locality_id]]$loggers[[logger_index]]$clean_info@step
+    }
     row_function <- function(tag, start, end, value){
         out_interval <- lubridate::intersect(lubridate::interval(start, end), date_interval)
         if(is.na(out_interval)) {
             return(list())
         }
-        start <- lubridate::floor_date(lubridate::int_start(out_interval), period)
-        end <- lubridate::floor_date(lubridate::int_end(out_interval), period)
+        if(is_agg){
+            start <- lubridate::floor_date(lubridate::int_start(out_interval), period)
+            end <- lubridate::floor_date(lubridate::int_end(out_interval), period)
+        } else {
+            start <- .states_floor_datetime(lubridate::int_start(out_interval), lubridate::int_start(date_interval), step)
+            end <- .states_floor_datetime(lubridate::int_end(out_interval), lubridate::int_start(date_interval), step)
+        }
         return(list(tag=tag,
                     start=start,
                     end=end,
@@ -312,34 +330,6 @@ mc_states_update <- function(data, states_table) {
         datetime <- data$localities[[locality_id]]$loggers[[logger_index]]$datetime
     }
     return(lubridate::interval(dplyr::first(datetime), dplyr::last(datetime)))
-}
-
-.states_update <- function(data_env, locality_id, logger_index, sensor_name, states_table) {
-    date_interval <- .states_get_item_range(data_env$data, locality_id, logger_index)
-    period <- .common_get_period_from_data(data_env$data, locality_id, logger_index)
-    row_function <- function(tag, start, end, value){
-        out_interval <- lubridate::intersect(lubridate::interval(start, end), date_interval)
-        if(is.na(out_interval)) {
-            return(list())
-        }
-        start <- lubridate::floor_date(lubridate::int_start(out_interval), period)
-        end <- lubridate::floor_date(lubridate::int_end(out_interval), period)
-        return(list(tag=tag,
-                    start=start,
-                    end=end,
-                    value=value))
-    }
-    states_table <- purrr::pmap_dfr(states_table, row_function)
-
-    if(.common_is_agg_format(data_env$data)) {
-        data_env$data$localities[[locality_id]]$sensors[[sensor_name]]$states <-
-            dplyr::bind_rows(data_env$data$localities[[locality_id]]$sensors[[sensor_name]]$states,
-                             states_table)
-    } else {
-        data_env$data$localities[[locality_id]]$loggers[[logger_index]]$sensors[[sensor_name]]$states <-
-            dplyr::bind_rows(data_env$data$localities[[locality_id]]$loggers[[logger_index]]$sensors[[sensor_name]]$states,
-                             states_table)
-    }
 }
 
 #' Delete sensor states (tags)
@@ -396,4 +386,22 @@ mc_states_delete <- function(data, localities=NULL, sensors=NULL, tags=NULL) {
     data$localities <- purrr::map(data$localities, locality_function)
 
     return(data)
+}
+
+.states_floor_sensor <- function(sensor, start_datetime, step) {
+    if(nrow(sensor$states) == 0) {
+        return(sensor)
+    }
+
+    sensor$states$start <- .states_floor_datetime(sensor$states$start, start_datetime, step)
+    sensor$states$end <- .states_floor_datetime(sensor$states$end, start_datetime, step)
+
+    return(sensor)
+}
+
+.states_floor_datetime <- function(datetime_values, start_datetime, step) {
+    start_seconds <- as.numeric(start_datetime)
+    datetime_seconds <- as.numeric(datetime_values) - start_seconds
+    result <- .common_as_utc_posixct(datetime_seconds %/% step * step + start_seconds)
+    return(result)
 }